clients/cli-client/requirements.txt
requests==2.32.3
python-socketio==5.11.0